Split UI building into several files
