+++ client/lib/main.dart
@@ -18,13 +18,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import 'package:blog_client/blog/commands.pb.dart';
 import 'package:blog_client/client.dart';
 import 'package:flutter/material.dart';
-import 'package:uuid/uuid.dart';
 
-import 'blog/blog.pb.dart';
-import 'blog/identifiers.pb.dart';
+import 'home_page.dart';
 
 void main() {
     var client = NetworkClient('http://localhost:4242', 'https://spine-dev.firebaseio.com/');
@@ -50,253 +47,3 @@ class BlogApp extends StatelessWidget {
     }
 }
 
-class BlogHomePage extends StatefulWidget {
-
-    BlogHomePage(this.title, this._client);
-
-    final String title;
-    final Client _client;
-
-    @override
-    _BlogPageState createState() => _BlogPageState(_client);
-}
-
-class _BlogPageState extends State<BlogHomePage> {
-
-    final Client _client;
-
-    Map<BlogId, Blog> blogs = {};
-    BlogView displayedBlog;
-
-    _BlogPageState(this._client);
-
-    void addBlog(Blog blog) {
-        setState(() {
-            blogs[blog.id] = blog;
-        });
-    }
-
-    void display(BlogView blog) {
-        setState(() {
-            displayedBlog = blog;
-        });
-    }
-
-    @override
-    void initState() {
-        super.initState();
-        _fetchBlogs();
-    }
-
-    void _fetchBlogs() {
-        _client.fetchBlogs().forEach(addBlog);
-    }
-
-    void _fetchBlogWithPosts(BlogId id) {
-        _client.fetchBlogWithPosts(id).then(display);
-    }
-
-    void _newBlog(String title) {
-        var id = BlogId()
-            ..uuid = Uuid().v4();
-        var command = CreateBlog()
-            ..id = id
-            ..title = title;
-        _client.post(command).then((value) => _fetchBlogs());
-    }
-
-    void _newPost(String title, String content) {
-        var id = PostId()
-            ..uuid = Uuid().v4();
-        var blogId = displayedBlog.id;
-        var command = CreatePost()
-            ..id = id
-            ..blog = blogId
-            ..title = title
-            ..body = content;
-        _client.post(command).then((value) => _publishPost(id, blogId));
-    }
-
-    void _publishPost(PostId post, BlogId blog) {
-        var command = PublishPost()
-            ..post = post
-            ..blog = blog;
-        _client.post(command).then((value) => _fetchBlogWithPosts(blog));
-    }
-
-    @override
-    Widget build(BuildContext context) {
-        return Scaffold(
-            appBar: AppBar(
-                title: Text(widget.title),
-            ),
-            floatingActionButton: Visibility(
-                child: FloatingActionButton(
-                    child: Icon(Icons.add),
-                    onPressed: () => showDialog(
-                        context: context,
-                        builder: _newPostDialog
-                    ),
-                ),
-                visible: displayedBlog != null,
-            ),
-            body: Row(
-                children: [
-                    Expanded(
-                        child: _blogList(context),
-                        flex: 1
-                    ),
-                    Expanded(
-                        child: Column(key: _blogKey(), children: _postList()),
-                        flex: 3
-                    )
-                ],
-            )
-        );
-    }
-
-    Key _blogKey() {
-        return displayedBlog != null ? _BlogKey(displayedBlog.id) : Key('No data');
-    }
-
-    List<Widget> _postList() {
-        return displayedBlog != null && displayedBlog.post.isNotEmpty
-            ? _bakePosts()
-            : [_empty('Nothing to see here')];
-    }
-
-    Widget _blogList(BuildContext context) {
-        var newBlogButton = Container(
-            color: Colors.redAccent,
-            child: TextButton(
-                child: Text(
-                    "NEW BLOG",
-                    style: TextStyle(color: Colors.white)
-                ),
-                onPressed: () {
-                    showDialog(
-                        context: context,
-                        builder: _newBlogDialog
-                    );
-                },
-        ));
-        Widget content;
-        if (blogs.isNotEmpty) {
-            var blogs = _bakeBlogs();
-            List<Widget> items = [];
-            items.add(newBlogButton);
-            items.addAll(blogs);
-            content = ListView(children: items);
-        } else {
-            content = newBlogButton;
-        }
-        var container = Container(
-            child: content,
-            color: Colors.white
-        );
-        return container;
-    }
-
-    Widget _newBlogDialog(BuildContext context) {
-        var titleController = TextEditingController();
-        return SimpleDialog(
-            title: Text("Create new blog..."),
-            children: [
-                Padding(
-                    child: TextField(
-                        decoration: InputDecoration(hintText: "My cool blog"),
-                        controller: titleController
-                    ),
-                    padding: EdgeInsets.all(8),
-                ),
-                TextButton(
-                    child: Text("CREATE"),
-                    onPressed: () {
-                        _newBlog(titleController.text);
-                        _closeDialog(context);
-                    },
-                )
-            ]
-        );
-    }
-
-    Widget _newPostDialog(BuildContext context) {
-        var titleController = TextEditingController();
-        var bodyController = TextEditingController();
-        return SimpleDialog(
-            children: [
-                TextField(
-                    decoration: InputDecoration(hintText: "New post"),
-                    controller: titleController,
-                ),
-                TextField(
-                    decoration: InputDecoration(hintText: "Write something"),
-                    controller: bodyController
-                ),
-                Row(children: [
-                    Expanded(child: TextButton(
-                        child: Text("CANCEL", textAlign: TextAlign.end),
-                        onPressed: () => _closeDialog(context),
-                    )),
-                    TextButton(
-                        child: Text("POST"),
-                        onPressed: () {
-                            _newPost(titleController.text, bodyController.text);
-                            _closeDialog(context);
-                        },
-                    )
-                ])
-            ]
-        );
-    }
-
-    _closeDialog(BuildContext context) =>
-        Navigator.of(context, rootNavigator: true).pop();
-
-    List<Widget> _bakeBlogs() => blogs.entries.map((blog) => GestureDetector(
-        onTap: () => _fetchBlogWithPosts(blog.key),
-        child: Padding(
-            child: Row(
-                key: _BlogKey(blog.key),
-                children: _displayBlog(blog.value)
-            ),
-            padding: EdgeInsets.only(bottom: 8),
-    ))).toList(growable: true);
-
-    List<Widget> _displayBlog(Blog blog) => [
-        Text(blog.title, textScaleFactor: 1.5),
-        Text('${blog.post.length} posts', textScaleFactor: 0.75)
-    ];
-
-    List<Widget> _bakePosts() => displayedBlog
-        .post
-        .expand(_displayPost)
-        .toList(growable: false);
-
-    List<Widget> _displayPost(PostItem post) => [
-        Text(post.title, textScaleFactor: 1.5, softWrap: true),
-        Text(post.body, softWrap: true)
-    ];
-
-    Widget _empty(String text) => Center(
-        child: Text(text, style: TextStyle(
-            color: Colors.blueGrey,
-            fontSize: 30
-        ))
-    );
-}
-
-class _BlogKey extends Key {
-
-    final BlogId id;
-
-    _BlogKey(this.id) : super.empty();
-
-    @override
-    bool operator ==(Object other) =>
-        identical(this, other) ||
-        other is _BlogKey && runtimeType == other.runtimeType && id == other.id;
-
-    @override
-    int get hashCode => id.hashCode;
-}
